sửa like

+++ src/controllers/vacation/index.js
@@ -355,16 +355,31 @@ const finishVacation = async (req, res) => {
 
 const deleteVacation = async (req, res) => {
   try {
-    const vacationId = req.params._id;
+    const vacationId = req.params.id;
+    
+
+    const existingVacation = await vacationModel.findById(vacationId);
+    if (!existingVacation) {
+      return res.status(404).json({ message: "Không tìm thấy kỳ nghỉ" });
+    }
 
     // Kiểm tra xem người đăng nhập có quyền xóa kỳ nghỉ không
     if (req.userId.toString() !== existingVacation.createdBy.toString()) {
       return res.status(403).json({
-        message: "Bạn không có quyền cập nhật kỳ nghỉ",
+        message: "Bạn không có quyền xóa kỳ nghỉ",
       });
     }
 
     const deletedVacation = await vacationModel.findByIdAndDelete(vacationId);
+
+    if (deletedVacation.milestones?.length != 0){
+      for (let i = 0; i < milestones.length; i++) {
+        const element = milestones[i];
+        const milestoneObj = milestoneModel.findById(element);
+      }
+    }
+
+    console.log(deletedVacation)
 
     return res
       .status(200)
